requirements.txt
flask
pyyaml
colorama
psutil
